requirements.txt
absl-py==1.0.0
altgraph==0.16.1      
astunparse==1.6.3     
attrs==22.1.0
autopep8==1.6.0       
beautifulsoup4==4.10.0
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.10
click==8.1.3
colorama==0.4.4
comtypes==1.1.14
cycler==0.11.0
Cython==0.29.28
distlib==0.3.4
filelock==3.4.2
Flask==2.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==2.5.1
flatbuffers==2.0
fonttools==4.28.3
frida==12.11.18
future==0.18.2
gast==0.4.0
google-auth==2.5.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.3
grpcio==1.43.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kaggle==1.5.12
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
lxml==4.7.1
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
mediapipe==0.8.11
numpy==1.21.5
oauthlib==3.1.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pefile==2019.4.18
Pillow==9.2.0
pipenv==2021.11.23
platformdirs==2.4.1
protobuf==3.19.3
psutil==5.9.2
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycaw==20220416
pycodestyle==2.8.0
pycparser==2.21
pydub==0.25.1
PyInstaller==3.5
PyJWT==2.5.0
pyparsing==3.0.6
pytesseract==0.3.9
python-dateutil==2.8.2
python-slugify==6.0.1
pytz==2021.3
pywin32-ctypes==0.2.0
PyYAML==5.3.1
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.2
scipy==1.5.3
six==1.16.0
SoundFile==0.10.3.post1
soupsieve==2.3.1
SpeechRecognition==3.8.1
SQLAlchemy==1.4.41
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
tesseract==0.1.3
text-unidecode==1.3
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.8
virtualenv==20.13.0
virtualenv-clone==0.5.7
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0